soonmeet_api/requirements.txt
requests
django>=2.2.4,<2.3.0
djangorestframework>=3.10.0,<3.11.0
psycopg2-binary>=2.8.3,<2.9.0
django-filter>=2.2.0,<2.3.0
django-cors-headers
django-admin-honeypot==1.1.0
